[Feat] Book 수정 - ID Data Type Long으로 변경 - Select * 컬럼 나열

+++ src/main/java/lgcns/eegoba/api/book/controller/BookController.java
@@ -2,8 +2,6 @@ package lgcns.eegoba.api.book.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.List;
-import lgcns.eegoba.api.base.vo.ApiResponseVO;
 import lgcns.eegoba.api.book.service.BookService;
 import lgcns.eegoba.api.book.vo.BookVO;
 import lgcns.eegoba.api.review.vo.ReviewVO;
@@ -14,6 +12,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.HttpStatusCodeException;
 
+import java.util.List;
+
 @Slf4j
 @RequiredArgsConstructor
 @RestController("bookController")
@@ -23,7 +23,7 @@ public class BookController {
   private final BookService bookService;
 
   @GetMapping(value = "/{bookId}")
-  public ApiResponseVO<Object> getBookById(@PathVariable(value = "bookId") Integer bookId)
+  public ApiResponseVO<Object> getBookById(@PathVariable(value = "bookId") Long bookId)
       throws HttpStatusCodeException {
     try {
       BookVO book = bookService.getBookById(bookId);
@@ -89,7 +89,7 @@ public class BookController {
 
   @PutMapping(value = "/update/{bookId}")
   public ApiResponseVO<Object> updateBook(
-      @PathVariable(value = "bookId") Integer bookId, @RequestBody BookVO bookVO)
+      @PathVariable(value = "bookId") Long bookId, @RequestBody BookVO bookVO)
       throws HttpStatusCodeException {
     try {
       if (bookService.getBookById(bookId) == null) {
@@ -115,7 +115,7 @@ public class BookController {
   }
 
   @PutMapping(value = "/review/{bookId}")
-  public ApiResponseVO<Object> getReviewListByBookId(@PathVariable(value = "bookId") Integer bookId)
+  public ApiResponseVO<Object> getReviewListByBookId(@PathVariable(value = "bookId") Long bookId)
       throws HttpStatusCodeException {
     try {
       if (bookService.getBookById(bookId) == null) {
